requirements.txt
asgiref==3.4.1
Django==3.2.5
django-admin-thumbnails==0.2.6
easy-colors==0.0.1
Pillow==8.3.1
python-decouple==3.4
pytz==2021.1
send-mail==0.0.2
six==1.16.0
sqlparse==0.4.1
whitenoise==5.0.1